% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPerformance.R
\name{aggregateMatrixList}
\alias{aggregateMatrixList}
\title{Apply a summary function to a list of matrixes of the same dimensions}
\usage{
aggregateMatrixList(
  l,
  func = function(x) {     mean(x, na.rm = T) },
  as.matrix = F,
  value.name = NULL
)
}
\arguments{
\item{l}{A list of matrices}

\item{func}{A summary function (default is mean())}

\item{as.matrix}{If TRUE, will return a single matrix with the aggregated values. If FALSE, will
return a melted dataframe.}

\item{value.name}{A custom column name for the value column of the melted dataframe}
}
\description{
Apply a summary function to a list of matrixes of the same dimensions
}
