% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{cacheAndReadData}
\alias{cacheAndReadData}
\title{Cache remote files and read as a dataframe}
\usage{
cacheAndReadData(
  remote.path,
  local.path = NULL,
  cache.dir = path.expand("~/Documents/R_cache/"),
  overwrite = F
)
}
\arguments{
\item{remote.path}{Remote path to txt or rds file}

\item{local.path}{Local path to copy to}

\item{cache.dir}{Path to the cache directory. If `local.path` is not specified, files will be
cached here.}

\item{overwrite}{Redownload data from remote?}
}
\value{
A dataframe
}
\description{
Cache remote files and read as a dataframe
}
