% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.confusionHeatmap.R
\name{calcFracCorrect}
\alias{calcFracCorrect}
\title{Calculate the fraction of correctly predicted samples}
\usage{
calcFracCorrect(actual, predicted, rm.non.pred.classes = T)
}
\arguments{
\item{actual}{A factor of the actual classes}

\item{predicted}{A factor of the predicted classes}

\item{rm.non.pred.classes}{Remove stats for classes that are not in the factor levels of
`predicted`}
}
\value{
A dataframe
}
\description{
Calculate the fraction of correctly predicted samples
}
