% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorizeFacetStrips.R
\name{colorizeFacetStrips}
\alias{colorizeFacetStrips}
\title{Add colors to facet strips of a ggplot object}
\usage{
colorizeFacetStrips(p, colors = NULL, return.gtable = T)
}
\arguments{
\item{p}{A ggplot}

\item{colors}{A character vector of colors}

\item{return.gtable}{If TRUE will return a gtable (use this when passing the plot to cowplot to
preserve plot alignments). If FALSE, the gtable will be converted to a ggplot using
`ggplotify::as.ggplot()`}
}
\value{
A gtable or ggplot object
}
\description{
Add colors to facet strips of a ggplot object
}
