% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predReport.merge.R
\name{combineLists}
\alias{combineLists}
\title{Combine multiple lists which have the same objects}
\usage{
combineLists(lists, exclude.objects = NULL, show.warnings = T, verbose = F)
}
\arguments{
\item{lists}{A list of lists, where each list has the same objects}

\item{exclude.objects}{A character vector specifying the names of the objects within each list to
exclude}

\item{show.warnings}{Show warnings?}

\item{verbose}{Show progress messages?}
}
\value{
A list
}
\description{
Combine multiple lists which have the same objects
}
