% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{dfToFeaturesAndResponse}
\alias{dfToFeaturesAndResponse}
\title{Prepares a feature dataframe for multiclass classifier training}
\usage{
dfToFeaturesAndResponse(df, colname.response = "response")
}
\arguments{
\item{df}{A dataframe}

\item{colname.response}{Name of the response column}
}
\value{
A list containing the feature matrix/dataframe, response vector, and response
one-hot encode matrix
}
\description{
Separates a dataframe containing features and response variable. The response
variable is also one hot encoded
}
