% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl-gpu.R
\name{dim.tbl_gpu}
\alias{dim.tbl_gpu}
\title{Get dimensions of a GPU table}
\usage{
\method{dim}{tbl_gpu}(x)
}
\arguments{
\item{x}{A \code{tbl_gpu} object.}
}
\value{
An integer vector of length 2: c(nrow, ncol).
Returns `c(NA, ncol)
}
\description{
Returns the number of rows and columns in a GPU table.
}
\examples{
if (has_gpu()) {



}
}
