% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featStatsPerClass.R
\name{featStatsPerClass}
\alias{featStatsPerClass}
\title{Summary stats on a dataframe of features for each class}
\usage{
featStatsPerClass(x, y, verbose = F)
}
\arguments{
\item{x}{A dataframe of features where rows are samples}

\item{y}{A character or factor vector of class labels for each sample in `x`}

\item{verbose}{Show progress messages?}
}
\value{
A long form dataframe with the min, max, and average feature values for each feature for
each class
}
\description{
This function calculates the average feature values between the target cancer type
samples vs other samples (interquartile mean for numeric data, or proportion for logical data),
as well as min and max feature values over all samples. This data is used for plotting feature
contributions in the patient report.
}
