% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{forceDfOrder}
\alias{forceDfOrder}
\title{Preserve order of dataframe values}
\usage{
forceDfOrder(df)
}
\arguments{
\item{df}{A dataframe}
}
\value{
A data frame with characters converted into factors
}
\description{
Converts character columns into factors to preserve the order of their values
}
