% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossValidateHpc.R
\name{gatherCvOutput}
\alias{gatherCvOutput}
\title{Gathers results from CV fold dirs and creates summary plots}
\usage{
gatherCvOutput(
  cv.out.dir,
  out.path,
  pattern = "^test_set_report.rds$",
  verbose = T
)
}
\arguments{
\item{cv.out.dir}{CV dir. Output path from `spawnCvJobs()`}

\item{out.path}{Path to write output}

\item{pattern}{Name/pattern of the rds file outputted by the training function}

\item{verbose}{Show progress messages?}
}
\value{
Writes merged CV data and plots to out.dir
}
\description{
Gathers results from CV fold dirs and creates summary plots
}
