% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{head2}
\alias{head2}
\title{Return first rows and columns of a matrix or dataframe}
\usage{
head2(m, rows = 10, cols = 10)
}
\arguments{
\item{m}{A matrix or dataframe}

\item{rows}{Number of rows}

\item{cols}{Number of cols}
}
\value{
The subsetted matrix or dataframe
}
\description{
Return first rows and columns of a matrix or dataframe
}
