% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probCal.main.R
\name{isoReg}
\alias{isoReg}
\title{Isotonic regression}
\usage{
isoReg(x, y)
}
\arguments{
\item{x}{x values}

\item{y}{y values}
}
\value{
A list which the objects:
- orig_data: Original x,y values ordered by the x values
- curve: The x,y values of the curve
}
\description{
A wrapper for stats::isoreg()
}
