% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probCal.main.R
\name{logReg}
\alias{logReg}
\title{Logistic regression (1 dimensional)}
\usage{
logReg(x, y)
}
\arguments{
\item{x}{Input x values}

\item{y}{Input y values}
}
\value{
A numeric vector with the intercept (b0) and gradient (b1)
}
\description{
A wrapper for stats::glm()
}
