% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl-gpu.R
\name{names.tbl_gpu}
\alias{names.tbl_gpu}
\title{Get column names of a GPU table}
\usage{
\method{names}{tbl_gpu}(x)
}
\arguments{
\item{x}{A \code{tbl_gpu} object.}
}
\value{
A character vector of column names.
}
\description{
Get column names of a GPU table
}
