% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmf.R
\name{nmfPerClass}
\alias{nmfPerClass}
\title{Decompose signatures separately for each class}
\usage{
nmfPerClass(
  A,
  response,
  nmf.args = list(),
  max.sig.dist = 0.1,
  multi.core = F,
  tmp.dir = NULL,
  verbose = T
)
}
\arguments{
\item{A}{A numeric matrix (rows=samples, columns=features)}

\item{response}{A character or factor vector indicating the class of each sample (i.e. row) in `A`}

\item{nmf.args}{A list of arguments that can be passed to `nmf()`}

\item{max.sig.dist}{Signatures will be hierarchical clustered based on pearson correlation
distance. The tree will be cut at height==`max.sig.dist`. Signature clusters below this height
(i.e. similar signatures) are considered to be redundant. From each cluster, the signature
decomposed from the class with the most samples will be kept, and the rest are discarded}

\item{multi.core}{If TRUE, will use multiple cores}

\item{tmp.dir}{A path a temporary directory, which if provided, enables resume capability}

\item{verbose}{Show progress messages? Can be 0,1,2}
}
\value{
A list containing the signature profile matrix (columns=feature, rows=signature) and MSE
values from the rank search
}
\description{
Decompose signatures separately for each class
}
