% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patientReport.main.R
\name{patientReport}
\alias{patientReport}
\title{Plot patient report}
\usage{
patientReport(
  probs = NULL,
  feat.contrib = NULL,
  sample.name = NULL,
  plot.title = sample.name,
  top.n.pred.classes = 3,
  top.n.features = 5,
  prob.thres.min = 0.1,
  prob.thres.rel.diff = 0.4,
  which.plots = c("probs", "feat.values"),
  rel.widths = c(1.4, 1)
)
}
\arguments{
\item{probs}{A numeric vector where the names are the predicted classes}

\item{feat.contrib}{A long form dataframe containing feature contribution data. This dataframe
contains the columns: sample, binary_rf, feature, contrib, feature_supplement}

\item{sample.name}{Name of the sample in `report` for which to make the patient report.
Required when input data is provided with `report`}

\item{plot.title}{Plot title}

\item{top.n.pred.classes}{The number of panels to display that show the details for the top
predicted classes (including the exact probabilities, and feature contributions).}

\item{top.n.features}{The number features to show in the feature contribution plots in the
details panels}

\item{prob.thres.min}{The feature contribution panels will be shown for predicted classes with a
probability higher than this threshold}

\item{prob.thres.rel.diff}{The feature contribution panels will be shown for the top predicted
classes, as long as the relative difference between the 1st and 2nd classes, 2nd and 3rd classes,
etc, are higher than this threshold. Note that `prob.thres.min` overrides this threshold}

\item{which.plots}{A character vector indicating which plots to show. Can be 'probs',
'feat.values', and/or 'feat.contribs'}

\item{rel.widths}{A numeric vector of the same length as `which.plots` specifying the relative
widths of the subplots}
}
\value{
A cowplot of ggplot objects
}
\description{
The patient report shows the predicted probabilities of each cancer type on a radar
plot, as well as the features contributing to the top predicted cancer types as bar plots
}
