% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.featImp.R
\name{topFeatures}
\alias{topFeatures}
\alias{maxImpPerFeatureType}
\title{Plot feature importance}
\usage{
topFeatures(
  m = NULL,
  top.n = 10,
  feature.type.colors = NULL,
  infer.feature.type = F,
  infer.feature.type.func = NULL,
  facet.ncol = NULL,
  facet.nrow = NULL,
  axis.x.breaks = NULL,
  hide.legend = F,
  as.list = F
)

maxImpPerFeatureType(m, infer.feature.type.func = NULL)
}
\arguments{
\item{m}{A numeric matrix where columns represent the features and rows represent the classes
(i.e. the names of each binary random forest)}

\item{top.n}{Top number of features to show}

\item{feature.type.colors}{A character vector of color hex codes with names being the feature
types}

\item{infer.feature.type}{Determine the feature type based on the tag/prefix. Everything before}

\item{infer.feature.type.func}{Function used to infer feature type. Defaults to
`function(x){ gsub('[.].+$','',x) }`, which takes the string before the first dot as the tag}

\item{facet.ncol}{Number of facet columns. If 'auto', this value will be
`round(sqrt(n_classes))`}

\item{facet.nrow}{Number of facet rows}

\item{axis.x.breaks}{A numeric vector indicating the x-axis breaks}

\item{hide.legend}{Hide legend?}

\item{as.list}{If FALSE (default), a faceted ggplot is returned (one facet for each class). If
TRUE, each class is plotted as separate ggplot and a list of these is returned}
}
\value{
A ggplot object
}
\description{
topFeatures(): Plot sorted feature importance barplots per class from a matrix of
feature importances. maxImpPerFeatureType(): Plot the max importance per feature type as a
heatmap
}
