% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPerformance.R
\name{plotFnr}
\alias{plotFnr}
\title{Plot false negative rate}
\usage{
plotFnr(actual, prob)
}
\arguments{
\item{actual}{A vector of the actual classes}

\item{prob}{A matrix of probabilities, where rows are the samples and columns are the cancer
types}
}
\value{
A ggplot object
}
\description{
The false negative rate can also be interpreted as the cumulative fraction of
samples per class that have a given probability.
}
