% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPerformance.R
\name{plotHeatmapFromMatrix}
\alias{plotHeatmapFromMatrix}
\title{Plot a heatmap from a matrix}
\usage{
plotHeatmapFromMatrix(
  m,
  x.lab = NULL,
  y.lab = NULL,
  legend.name = "value",
  show.labels = F,
  custom.labels = NULL,
  palette = "YlGnBu",
  palette.direction = -1,
  invert.y = F,
  x.title.position = "bottom"
)
}
\arguments{
\item{m}{A numeric matrix}

\item{x.lab}{x-axis labels}

\item{y.lab}{y-axis labels}

\item{legend.name}{Legend title}

\item{show.labels}{Show raw values within each cell?}

\item{palette}{Name of the brewer color palette}

\item{palette.direction}{Color palette direction. Can be 1 (forward) or -1 (reverse).}

\item{invert.y}{Invert the y-axis?}

\item{x.title.position}{Position of the x axis title}
}
\value{
A ggplot object
}
\description{
Plot a heatmap from a matrix
}
