% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotPerfHeatmap}
\alias{plotPerfHeatmap}
\title{Plot heatmap of classifier performance}
\usage{
plotPerfHeatmap(
  actual = NULL,
  predicted = NULL,
  cv_out = NULL,
  rel.heights = c(0.3, 1),
  sort = F,
  rel.values = T,
  metrics = c("f1", "prec", "tpr"),
  show.weighted.mean = T
)
}
\arguments{
\item{actual}{A vector of the actual classes}

\item{predicted}{A vector of the predicted classes}

\item{cv_out}{Alternative input to `actual` and `predicted`. Cross validation output in the
form of a list that has 'imp' object (e.g. cv_out[[1]]$test_set$actual and
cv_out[[1]]$test_set$predicted)}

\item{rel.heights}{A numeric vector of length 2. Relative heights of the upper and lower plots}

\item{sort}{Sort cancer type by number of % sample of correctly classified}

\item{rel.values}{In lower plot, show absolute number or %}

\item{metrics}{A character vector indicating which performance metrics to show in the upper plot.
See documentation for mltoolkit to see which metrics are available}

\item{show.weighted.mean}{In upper plot, calculated the weighted mean or normal mean?}
}
\value{
A cowplot object
}
\description{
Creates two plots. Upper plot shows performance metrics per cancer type and lower
plot shows the number or % of samples correctly/incorrectly classified
}
