% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPerformance.R
\name{plotTopFeatures}
\alias{plotTopFeatures}
\title{Make barplots from a matrix of feature importances (for multiclass classification)}
\usage{
plotTopFeatures(
  m = NULL,
  cv_out = NULL,
  top.n = 10,
  n.row = NULL,
  n.col = NULL,
  feature.type.colors = NULL,
  infer.feature.type = F,
  infer.feature.type.func = NULL
)
}
\arguments{
\item{m}{A numeric matrix where columns represent the features and rows represent the labels}

\item{cv_out}{Alternative input to `m`. Cross validation output in the form of a list that has
'imp' object (e.g. cv_out[[1]]$imp)}

\item{top.n}{Top number of features to show}

\item{n.row}{Number of facet rows}

\item{n.col}{Number of facet columns}

\item{feature.type.colors}{A character vector of color hex codes with names being the feature
types}

\item{infer.feature.type}{Determine the feature type based on the tag/prefix. Everything before
the first dot (.) is considered the feature tag}
}
\value{
A ggplot object
}
\description{
Make barplots from a matrix of feature importances (for multiclass classification)
}
