% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predReport.summary.R
\name{predReportSummary}
\alias{predReportSummary}
\title{Summarize prediction report}
\usage{
predReportSummary(
  report,
  show.which.top.classes = NULL,
  show.top.class.probs = T,
  show.which.top.features = NULL,
  show.top.features.contribs = T
)
}
\arguments{
\item{report}{A list with the objects with the names: probs_raw, prob,
class_pred, class_actual, feat_contrib, imp}

\item{show.which.top.classes}{An integer vector specifying the top n predicted classes to show}

\item{show.which.top.features}{An integer vector specifying the top features contributing to the
top predicted class to show}
}
\value{
A dataframe
}
\description{
Summarize prediction report
}
