% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probCal.main.R
\name{probCalCurves}
\alias{probCalCurves}
\alias{probCal}
\title{Generate probability calibration curves}
\usage{
probCalCurves(
  actual = NULL,
  probs = NULL,
  report = NULL,
  output = c("curve", "plot", "plotdata"),
  method = c("isotonic", "logistic")
)

probCal(probs, cal.curves)
}
\arguments{
\item{actual}{A vector of the actual classes}

\item{probs}{A matrix where rows are samples, cols are binary random forest names, and cells are
the prediction probabilities from each random forest}

\item{report}{A list with the objects with the names: prob, class_actual}

\item{output}{If 'coords' returns a dataframe with the x,y values of the calibration curve for
each `actual` class. If 'plot', returns a ggplot object. If 'plotdata', returns a list with the
raw plot data}

\item{method}{Type of regression to perform. Can be 'isotonic' or 'logistic'.}

\item{cal.curves}{The output from `probCalCurves(..., output='curve')`}
}
\value{
`probCalCurves()`: A dataframe or ggplot object (see `output`). `probCal()` a matrix
of calibrated probabilities per class
}
\description{
`probCalCurves()` generates calibration curves to map raw classifier
(pseudo-)probabilibities to real probabilities. This is done using isotonic regression, with
x-values being raw probs and y-values being the sample labels (1 or 0). `probCal()` is used to
map raw probabilities to calibrated probabilities
}
