% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probCal.reliabilityPlot.R
\name{reliabilityStats}
\alias{reliabilityStats}
\alias{reliabilityPlot}
\title{Reliability plot}
\usage{
reliabilityStats(
  actual = NULL,
  probs = NULL,
  report = NULL,
  n.bins = 10,
  bootstrap.iters = 50,
  bootstrap.prop = 0.66,
  seed = 1,
  verbose = F
)

reliabilityPlot(
  actual = NULL,
  probs = NULL,
  report = NULL,
  stats = NULL,
  min.samples.per.bin = 5,
  show.labels = T,
  facet.nrow = NULL,
  facet.ncol = NULL,
  ...
)
}
\arguments{
\item{actual}{A vector of the actual classes}

\item{probs}{A matrix where rows are samples, cols are binary random forest names, and cells are
the prediction probabilities from each random forest}

\item{report}{A list with the objects with the names: prob, class_actual}

\item{n.bins}{Number of (equidistant bins)}

\item{bootstrap.iters}{Number of bootstrap iterations to calculate error bars (standard deviation)}

\item{bootstrap.prop}{Proportion of all samples to sample from for boostrapping}

\item{seed}{Bootstrap seed}

\item{verbose}{Show progress messages?}

\item{stats}{The output of `reliabilityStats()`}

\item{min.samples.per.bin}{Points with fewer samples than this will no be considered for the
overlaid (filtered) reliability curve}

\item{show.labels}{Show labels at each point}

\item{facet.nrow}{Number of facet rows in plot}

\item{facet.ncol}{Number of facet columns in plot}
}
\value{
For `reliabilityStats()` a dataframe. For `reliabilityPlot()` a ggplot
}
\description{
Bin probabilities and calculate the proportion of positive class samples per bin.
The proportion of positive class samples (y-values) will be plotted against the midpoint of each
probability bin (x-values), and a curve will connect the points. A 'filtered' curve is also shown
where points supported by too few samples (<`min.samples.per.bin`) will be ignored for plotting
this curve.
}
