% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balanceClasses.R
\name{resampleClasses}
\alias{resampleClasses}
\title{Resolve class imbalances by simple resampling}
\usage{
resampleClasses(
  df,
  colname.response = "response",
  resample.ratios = NULL,
  target.sample.sizes = NULL,
  return.data = T
)
}
\arguments{
\item{df}{A dataframe containing the features and response variable}

\item{colname.response}{Column name of the response variable}

\item{resample.ratios}{A named numeric vector indicating the ratio of upsampling for each class.
Values < 1 indicate downsampling. Values == 1 indicates no resampling}

\item{target.sample.sizes}{A named integer vector indicating the target number of samples for
each class}

\item{return.data}{If TRUE, returns the resampled dataframe. If FALSE, returns the row indexes.}
}
\value{
A resampled dataframe or a vector of row indexes
}
\description{
Resolve class imbalances by simple resampling
}
