% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/training.R
\name{rmColumns}
\alias{rmColumns}
\title{Remove columns by name}
\usage{
rmColumns(df, columns, drop = F)
}
\arguments{
\item{df}{A dataframe or matrix}

\item{columns}{Column names}

\item{drop}{If TRUE the result is coerced to the lowest possible dimension}
}
\value{
A dataframe or matrix
}
\description{
Remove columns by name
}
