% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predReport.summary.R
\name{summary.predReport}
\alias{summary.predReport}
\title{Summarize prediction report}
\usage{
\method{summary}{predReport}(
  report,
  top.n.classes = NULL,
  show.class.probs = T,
  simplify.pred.values = F,
  prob.type = NULL,
  top.n.feat = NULL,
  top.n.feat.classes = 1,
  show.feat.contribs = T
)
}
\arguments{
\item{report}{A list with the objects with the names: probs_raw, prob,
class_pred, class_actual, feat_contrib, imp}

\item{top.n.classes}{An integer specifying the top n predicted classes to show}

\item{show.class.probs}{When `top.n.classes`==TRUE, also show the predicted
probabilities?}

\item{simplify.pred.values}{If TRUE, the top n predicted classes and probs will be combined as a
string in the form {class}={prob}}

\item{prob.type}{Can be 'prob' (raw probabilities) or 'prob_scaled' (scaled probabilities)}

\item{top.n.feat}{An integer specifying the top features contributing to the top predicted
class to show}

\item{top.n.feat.classes}{An integer specifying the number of top classes for which to show the
top contributing features}

\item{show.feat.contribs}{When `top.n.feat`==TRUE, also show the feature
contributions?}
}
\value{
A dataframe
}
\description{
Summarize prediction report
}
