% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.targetClassProb.R
\name{targetClassProb}
\alias{targetClassProb}
\title{Plot probability of target class}
\usage{
targetClassProb(
  actual = NULL,
  probs = NULL,
  report = NULL,
  output = c("plot.sorted_probs", "plot.per_cutoff", "raw.sorted_probs",
    "raw.per_cutoff"),
  show.labels = F,
  show.bars = F,
  cutoffs = seq(0, 1, 0.1),
  force.discrete.x.axis = F,
  show.all.classes.stats = if (show.bars) {     F } else {     T }
)
}
\arguments{
\item{actual}{A vector of the actual classes}

\item{probs}{A matrix where rows are samples, cols are binary random forest names, and cells are
the prediction probabilities from each random forest}

\item{report}{A list with the objects with the names: prob, class_actual}

\item{output}{If 'plot', returns a ggplot object, else 'raw' returns the raw data}

\item{show.labels}{Show labels indicating the numerical values at each point?}

\item{show.bars}{For each sample, show bars colored by the predicted class? Only applies when
`output`=='plot.sorted_probs'}

\item{cutoffs}{A numeric vector of cutoffs. Applies only when `plot.type`='per_cutoff'}

\item{force.discrete.x.axis}{Force discrete x axis}

\item{show.all.classes.stats}{Show the panel with the stats for all samples (i.e. not split by
actual class)}
}
\value{
A ggplot object
}
\description{
For samples within each actual class, plot the probability of said actual class.
The type of plot returned depends on `output`:
  'plot.sorted_probs'  x: cutoff, y: probability of correct prediction
  'plot.per_cutoff'    x: cutoff, y: fraction of samples >= cutoff
}
