% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.topPredClasses.R
\name{topPredClasses}
\alias{topPredClasses}
\title{Plot top predicted classes}
\usage{
topPredClasses(
  actual,
  predicted = NULL,
  probs = NULL,
  top.n = 2,
  output = c("plot", "raw"),
  show.all.classes.stats = T,
  sort.classes = F,
  show.labels = F
)
}
\arguments{
\item{actual}{A factor vector of the actual classes}

\item{predicted}{A factor vector of the predicted classes}

\item{top.n}{The top predicted classes to show}

\item{output}{If 'plot', returns a ggplot object, else 'raw' returns the raw data}

\item{show.all.classes.stats}{Show stats for all samples (i.e. not split by actual class)}

\item{sort.classes}{Sort classes by the predicted class?}

\item{show.labels}{Show labels indicating the numerical values at each point?}
}
\value{
A ggplot object
}
\description{
Plot top predicted classes
}
