% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomForestEnsemble.train.R
\name{trainRandomForest}
\alias{trainRandomForest}
\title{Train a random forest (for binary classification)}
\usage{
trainRandomForest(
  train,
  test = NULL,
  colname.response = "response",
  do.feat.sel = T,
  feat.sel.alternative = NULL,
  feat.sel.max.pvalue = 0.01,
  feat.sel.min.cliff.delta = 0.1,
  feat.sel.min.cramer.v = 0.1,
  feat.sel.top.n.features = NULL,
  feat.sel.whitelist = NULL,
  balance.classes = c("none", "resample", "class_weights"),
  k.inner = 5,
  n.resamples.true = 5,
  n.resamples.false = 5,
  midpoint.type = "geometric",
  min.size.diff = 15,
  max.upsample.ratio = 5,
  max.upsample.size = 300,
  min.downsample.ratio = NULL,
  min.downsample.size = 1000,
  max.imbalance.ratio = 15,
  max.pairs = 10,
  ntree = 500,
  get.local.increments = T,
  feat.tmp.path = NULL,
  model.tmp.path = NULL,
  seed = NULL,
  msg.prefix = NULL,
  verbose = 1
)
}
\arguments{
\item{train}{A dataframe containing the training features and response column.}

\item{test}{A dataframe containing the training features and response column. Used for testing
the performance of the trained model}

\item{colname.response}{The column name of the response variable (i.e. training labels)}

\item{do.feat.sel}{Perform univariate feature selection?}

\item{feat.sel.alternative}{Can be 'two.sided', 'greater' or 'less'. Can be provided as a
single string, or a character vector of the same length as the number of features}

\item{feat.sel.max.pvalue}{Only features with p-value lower than this will be kept}

\item{feat.sel.min.cliff.delta}{Cliff delta (effect size) threshold for keeping continous features}

\item{feat.sel.min.cramer.v}{Cramer's V (effect size) threshold for keeping categorical features}

\item{feat.sel.top.n.features}{Only keep the top number of features}

\item{feat.sel.whitelist}{A character vector of feature names in x to keep (i.e. ignore feature
selection for these features)}

\item{balance.classes}{If 'resample', up and downsampling will be performed. Resampling ratios
are determined by an inner cross-validation. If 'class_weights', class weights equal to
1/n_samples_in_class will be used. If 'none', no class balancing will be performed.}

\item{k.inner}{Number of inner cross-validation folds to use to determine resampling ratios}

\item{n.resamples.true}{Number of resampling values to generate for the TRUE class}

\item{n.resamples.false}{Number of resampling values to generate for the FALSE class}

\item{midpoint.type}{Can be 'geometric', 'arithmetic', or 'none'. Calculate the resampling values
from a->midpoint and b->midpoint? If 'none', resampling values will be calculated from a->b and
b->a.}

\item{min.size.diff}{Default=30. Minimum difference between resampling values (integer).}

\item{max.upsample.ratio}{See documentation for `resamplingGrid()`}

\item{max.upsample.size}{See documentation for `resamplingGrid()`}

\item{min.downsample.ratio}{See documentation for `resamplingGrid()`}

\item{min.downsample.size}{See documentation for `resamplingGrid()`}

\item{max.imbalance.ratio}{See documentation for `resamplingGrid()`}

\item{max.pairs}{See documentation for `resamplingGrid()`}

\item{ntree}{Number of decision trees in the random forest}

\item{get.local.increments}{If TRUE, get local increments to calculate feature contributions}

\item{feat.tmp.path}{Tmp path for feature selection output}

\item{model.tmp.path}{Tmp path for trained random forest}

\item{seed}{Random seed as an integer}

\item{msg.prefix}{Add prefix to progress messages. If not specified, no prefix will be displayed.}

\item{verbose}{Show progress? Can be 0, 1, 2 (increasing verbosity)}
}
\value{
A list containing the training output
}
\description{
Train a random forest (for binary classification)
}
