% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/training.R
\name{trainRandomForestEnsemble}
\alias{trainRandomForestEnsemble}
\title{Train a random forest ensemble (for multiclass classification)}
\usage{
trainRandomForestEnsemble(
  train,
  test = NULL,
  colname.response = "response",
  do.rmd.nmf = F,
  args.trainRandomForest = list(),
  inner.holdout.fraction = c(1, 3),
  seed = NULL,
  msg.prefix = NULL,
  tmp.dir = NULL,
  rm.tmp.dir = T,
  multi.core = F,
  verbose = 1
)
}
\arguments{
\item{train}{A dataframe containing the training features and response column.}

\item{test}{A dataframe containing the training features and response column. Used for testing
the performance of the trained model}

\item{colname.response}{The column name of the response variable (i.e. training labels)}

\item{do.rmd.nmf}{Perform NMF to convert RMD bins to RMD signatures?}

\item{args.trainRandomForest}{A list of arguments that can be passed to `trainRandomForest()`}

\item{seed}{Random seed as an integer}

\item{tmp.dir}{A path a temporary directory, which if provided, enables resume capability}

\item{rm.tmp.dir}{If TRUE will remove the tmp dir if training completes successfully}

\item{multi.core}{Use multiple cores?}

\item{verbose}{Show progress? Can be 0, 1, 2 (increasing verbosity)}

\item{fold.num}{Cross validation fold number. Only used in the prefix of progress messages. If
not specified, no prefix will be displayed.}
}
\value{
A list containing the training output
}
\description{
Train a random forest ensemble (for multiclass classification)
}
