% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probCal.main.R
\name{isoReg}
\alias{isoReg}
\title{Isotonic regression}
\usage{
isoReg(x, y, na.rm = F)
}
\arguments{
\item{x}{Input x values}

\item{y}{Input y values}

\item{na.rm}{Remove pairs of x and y values that have NA}
}
\value{
A dataframe containing the x,y values of the curve
}
\description{
A wrapper for stats::isoreg()
}
