% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patientReport.main.R
\name{patientReport}
\alias{patientReport}
\title{Plot patient report}
\usage{
patientReport(
  probs = NULL,
  feat.contrib = NULL,
  sample.name = NULL,
  plot.title = sample.name,
  top.n.pred.classes = 3,
  top.n.features = 5,
  max.rel.diff = 0.6,
  rel.widths = c(2, 1)
)
}
\arguments{
\item{probs}{A numeric vector where the names are the predicted classes}

\item{feat.contrib}{A long form dataframe containing feature contribution data. This dataframe
contains the columns: sample, binary_rf, feature, contrib, feature_supplement}

\item{sample.name}{Name of the sample in `report` for which to make the patient report.
Required when input data is provided with `report`}

\item{plot.title}{Plot title}

\item{top.n.pred.classes}{The number of panels to display that show the details for the top
predicted classes (including the exact probabilities, and feature contributions).}

\item{top.n.features}{The number features to show in the feature contribution plots in the
details pnaels}

\item{max.rel.diff}{}

\item{rel.widths}{A numeric vector of length 2 specifying the relative widths of the probability
plot and feature contribution plots}
}
\value{
A ggplot grob
}
\description{
The patient report shows the predicted probabilities of each cancer type on a radar
plot, as well as the features contributing to the top predicted cancer types as bar plots
}
