% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPerfHeatmap.R
\name{plotPerfHeatmap}
\alias{plotPerfHeatmap}
\title{Plot heatmap of classifier performance}
\usage{
plotPerfHeatmap(
  actual = NULL,
  predicted = NULL,
  rel.heights = c(perf = 0.3, counts = 0.15, heatmap = 1),
  sort.classes = F,
  rel.values = T,
  predicted.classes.only = F,
  metrics = c("f1", "prec", "tpr"),
  show.weighted.mean = T
)
}
\arguments{
\item{actual}{A factor vector of the actual classes}

\item{predicted}{A factor vector of the predicted classes}

\item{rel.heights}{A numeric vector of length 3. Relative heights of the plots}

\item{sort.classes}{Sort cancer type by number of % sample of correctly classified}

\item{rel.values}{In lower plot, show absolute number or %}

\item{predicted.classes.only}{Only show classes that are present in predicted}

\item{metrics}{A character vector indicating which performance metrics to show in the upper plot.
See documentation for mltoolkit to see which metrics are available}

\item{show.weighted.mean}{In upper plot, calculated the weighted mean or normal mean?}
}
\value{
A cowplot object
}
\description{
Creates two plots. Upper plot shows performance metrics per cancer type and lower
plot shows the number or % of samples correctly/incorrectly classified
}
