% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomForestEnsemble.predict.R
\name{predict.randomForestEnsemble}
\alias{predict.randomForestEnsemble}
\title{Predict method for random forest ensemble}
\usage{
\method{predict}{randomForestEnsemble}(
  object,
  newdata,
  type = "report",
  gender.feature.name = "gender.gender",
  classes.female = c("Cervix", "Ovary", "Uterus"),
  classes.male = "Prostate",
  calc.feat.contrib = T,
  top.n.pred.classes = NULL,
  top.n.features = 5,
  verbose = F
)
}
\arguments{
\item{object}{An object of class randomForestEnsemble}

\item{newdata}{A data frame or matrix containing new data. (Note: If not given, the out-of-bag
prediction in object is returned)}

\item{type}{'class', 'prob' or 'votes', indicating the type of output: predicted values,
matrix of class probabilities, or matrix of vote counts.}

\item{gender.feature.name}{The name of the feature specifying the gender of each sample}

\item{classes.female}{A character vector of female cancer type classes. For male samples (as
determined by `gender.feature.name`), the probabilities outputted by the binary random forests
corresponding to `classes.female` will be set to zero}

\item{classes.male}{A character vector of male cancer type classes. Similar to `classes.female`}

\item{calc.feat.contrib}{Calculate feature contributions?}

\item{top.n.pred.classes}{Number of top predicted classes (i.e. binary random forest names) to
keep features from when calculating feature contributions}

\item{top.n.features}{Number of top features for each sample to keep when calculating feature
contributions}

\item{verbose}{Show progress messages? Can be 0,1,2}
}
\value{
'class': predicted classes (the classes with majority vote).
'prob': matrix of class
probabilities (one column for each class and one row for each input)
'report': a list containing the following objects: probs_raw, probs_adjusted, responses_pred,
feat_contrib
}
\description{
Predict method for random forest ensemble
}
