% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/training.R
\name{predict.randomForestEnsemble}
\alias{predict.randomForestEnsemble}
\title{Predict method for random forest ensemble}
\usage{
\method{predict}{randomForestEnsemble}(
  object,
  newdata,
  type = "response",
  gender.feature.name = "purple.gender",
  classes.female = c("Cervix", "Ovary", "Uterus"),
  classes.male = "Prostate",
  verbose = F
)
}
\arguments{
\item{object}{An object of class randomForestEnsemble}

\item{newdata}{A data frame or matrix containing new data. (Note: If not given, the out-of-bag
prediction in object is returned)}

\item{type}{'response', 'prob' or 'votes', indicating the type of output: predicted values,
matrix of class probabilities, or matrix of vote counts. 'class' is allowed, but automatically
converted to "response", for backward compatibility.}
}
\value{
'response': predicted classes (the classes with majority vote). 'prob' matrix of class
probabilities (one column for each class and one row for each input). 'vote'
matrix of vote counts (one column for each class and one row for each new input); either in raw
counts or in fractions (if norm.votes=TRUE).
}
\description{
Predict method for random forest ensemble
}
