% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balanceClasses.R
\name{resamplingGrid}
\alias{resamplingGrid}
\title{Create pairs of resample sizes/ratios for performing a grid search}
\usage{
resamplingGrid(
  a,
  b,
  breaks.a = 4,
  breaks.b = 4,
  midpoint.type = "geometric",
  min.size.diff = 30,
  max.upsample.ratio = 10
)
}
\arguments{
\item{a}{The sample size of cohort 1}

\item{b}{The sample size of cohort 1}

\item{breaks.a}{Number of resampling values to generate for a}

\item{breaks.b}{Number of resampling values to generate for b}

\item{midpoint.type}{Can be 'geometric', 'arithmetic', or 'none'. Calculate the resampling values
from a->midpoint and b->midpoint? If 'none', resampling values will be calculated from a->b and
b->a.}

\item{min.size.diff}{Default=30. Minimum difference between resampling values (integer).}

\item{max.upsample.ratio}{Default=10. Remove pairs where a or b are upsampled more than this
value}
}
\value{
A dataframe of target sample sizes and resampling ratios for a and b
}
\description{
Create pairs of resample sizes/ratios for performing a grid search
}
