% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossValidateHpc.R
\name{spawnCvJobs}
\alias{spawnCvJobs}
\title{Spawn CV jobs for submission with SLURM}
\usage{
spawnCvJobs(
  train.data.path = NULL,
  df = NULL,
  train.script.path,
  cv.out.dir = paste0(dirname(train.script.path), "/cv_out/"),
  colname.response = "response",
  k = 20,
  seed = 1,
  rm.path.prefix = "/Users/lnguyen",
  time = "2:00:00",
  mem = "16G",
  verbose = T
)
}
\arguments{
\item{train.data.path}{Path to rds file. A dataframe of features and the response variable}

\item{df}{A dataframe of features and the response variable}

\item{train.script.path}{Path to training R script. The following input arguments are required in
within the script:
args <- commandArgs(trailingOnly=T)
training_data <- args[1]
fold_indexes <- args[2]
out_path <- args[3]
seed <- args[4]}

\item{cv.out.dir}{CV output dir. Jobs, input data and output data will be stored here}

\item{colname.response}{The column name of the response variable (i.e. training labels).}

\item{k}{Number of cross validation folds.}

\item{seed}{Random seed as in integer.}

\item{rm.path.prefix}{Prefix to remove from all paths in the job script}

\item{verbose}{Show progress messages?}
}
\value{
Writes jobs to cv.out.dir
}
\description{
Spawn CV jobs for submission with SLURM
}
