% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predReport.summary.R
\name{summary.predReport}
\alias{summary.predReport}
\title{Summarize prediction report}
\usage{
\method{summary}{predReport}(
  report,
  top.n.classes = NULL,
  show.class.probs = T,
  simplify.pred.values = F,
  prob.type = NULL,
  top.n.feat = NULL,
  show.feat.contribs = T,
  object.names = NULL
)
}
\arguments{
\item{report}{A list with the objects with the names: probs_raw, prob,
class_pred, class_actual, feat_contrib, imp}

\item{top.n.classes}{An integer specifying the top n predicted classes to show}

\item{show.class.probs}{When `top.n.classes`==TRUE, also show the predicted
probabilities?}

\item{simplify.pred.values}{If TRUE, the top n predicted classes and probs will be combined as a
string in the form {class}={prob}}

\item{prob.type}{Can be 'prob' (raw probabilities) or 'prob_scaled' (scaled probabilities)}

\item{top.n.feat}{An integer specifying the top features contributing to the top predicted
class to show}

\item{show.feat.contribs}{When `top.n.feat`==TRUE, also show the feature
contributions?}

\item{object.names}{Specify custom object names within `report` with a named character vector in
the form: `c(prob='prob',class_actual='class_actual',feat_contrib='feat_contrib')`.
The vector values are the custom object names, and the vector names correspond to the names used
internally in the function. `object.names` can be NULL, or one or more named values}
}
\value{
A dataframe
}
\description{
Summarize prediction report
}
