% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.topnAcc.R
\name{topnAcc}
\alias{topnAcc}
\title{Plot top-n accuracy}
\usage{
topnAcc(
  report = NULL,
  actual = NULL,
  probs = NULL,
  top.n = 3,
  output = c("plot", "raw"),
  show.all.classes.stats = T,
  max.prob = NULL
)
}
\arguments{
\item{report}{A list with the objects with the names: prob, class_actual}

\item{actual}{A vector of the actual classes}

\item{probs}{A matrix where rows are samples, cols are binary random forest names, and cells are
the prediction probabilities from each random forest}

\item{top.n}{Number of top classes to show stats for}

\item{output}{If 'plot', returns a ggplot object, else 'raw' returns the raw data}

\item{show.all.classes.stats}{Show the panel with the stats for all samples (i.e. not split by
actual class)}

\item{max.prob}{Filter for samples <= than this probability}
}
\value{
A ggplot object
}
\description{
For samples within each actual class, plot the top-n accuracy of said actual class.
}
