% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/training.R
\name{trainRandomForest}
\alias{trainRandomForest}
\title{Train a random forest (for binary classification)}
\usage{
trainRandomForest(
  train,
  test = NULL,
  colname.response = "response",
  do.feat.sel = T,
  feat.sel.v.alternative = NULL,
  feat.sel.max.pvalue = 0.01,
  feat.sel.min.effect.size = 3,
  feat.sel.top.n.features = NULL,
  balance.classes = F,
  k.inner = 5,
  n.resamples.true = 4,
  n.resamples.false = 4,
  midpoint.type = "geometric",
  min.size.diff = NULL,
  max.upsample.ratio = 10,
  ntree = 500,
  get.local.increments = T,
  model.tmp.path = NULL,
  seed = NULL,
  msg.prefix = NULL,
  verbose = 1
)
}
\arguments{
\item{train}{A dataframe containing the training features and response column.}

\item{test}{A dataframe containing the training features and response column. Used for testing
the performance of the trained model}

\item{colname.response}{The column name of the response variable (i.e. training labels)}

\item{do.feat.sel}{Perform univariate feature selection?}

\item{feat.sel.max.pvalue}{Only features with p-value lower than this will be kept}

\item{feat.sel.min.effect.size}{Effect size threshold for keeping features. Default: 3
(applies to both +3 and -3 effect sizes)}

\item{feat.sel.top.n.features}{Only keep the top number of features}

\item{n.resamples.true}{Number of resampling values to generate for the TRUE class}

\item{n.resamples.false}{Number of resampling values to generate for the FALSE class}

\item{midpoint.type}{Can be 'geometric', 'arithmetic', or 'none'. Calculate the resampling values
from a->midpoint and b->midpoint? If 'none', resampling values will be calculated from a->b and
b->a.}

\item{min.size.diff}{Default=30. Minimum difference between resampling values (integer).}

\item{max.upsample.ratio}{Default=10. Remove pairs where a or b are upsampled more than this
value}

\item{ntree}{Number of decision trees in the random forest}

\item{get.local.increments}{If TRUE, get local increments to calculate feature contributions}

\item{seed}{Random seed as an integer}

\item{msg.prefix}{Add prefix to progress messages. If not specified, no prefix will be displayed.}

\item{verbose}{Show progress? Can be 0, 1, 2 (increasing verbosity)}
}
\value{
A list containing the training output
}
\description{
Train a random forest (for binary classification)
}
