% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/training.R
\name{trainRandomForestEnsemble}
\alias{trainRandomForestEnsemble}
\title{Train a random forest ensemble (for multiclass classification)}
\usage{
trainRandomForestEnsemble(
  train,
  test = NULL,
  colname.response = "response",
  do.feat.sel = TRUE,
  feat.sel.max.qvalue = 0.01,
  feat.sel.max.pvalue = 1,
  feat.sel.top.n.features = NULL,
  calc.imp = T,
  imp.metric = "mda",
  ntree = 500,
  get.local.increments = TRUE,
  seed = NULL,
  fold.num = NULL,
  verbose = 1
)
}
\arguments{
\item{train}{A dataframe containing the training features and response column.}

\item{test}{A dataframe containing the training features and response column. Used for testing
the performance of the trained model}

\item{colname.response}{The column name of the response variable (i.e. training labels)}

\item{do.feat.sel}{Perform univariate feature selection?}

\item{feat.sel.max.qvalue}{Only features with q-value (from univariate feature selection) lower
than this will be kept}

\item{feat.sel.max.pvalue}{Only features with p-value (from univariate feature selection) lower
than this will be kept}

\item{feat.sel.top.n.features}{Only keep the top number of features}

\item{calc.imp}{Calculate feature importance?}

\item{imp.metric}{Which type of feature importance to calculate}

\item{ntree}{Number of decision trees in the random forest}

\item{get.local.increments}{If TRUE, get local increments to calculate feature contributions}

\item{seed}{Random seed as an integer}

\item{fold.num}{Cross validation fold number. Only used in the prefix of progress messages. If
not specified, no prefix will be displayed.}

\item{verbose}{Show progress? Can be 0, 1, 2 (increasing verbosity)}
}
\value{
A list containing the training output
}
\description{
Train a random forest ensemble (for multiclass classification)
}
